dev-requirements.txt
-r requirements.txt

# parsing files
pyflakes==3.0.1
pylint==2.16.4
pylama==8.4.1
tox==4.4.6

# unit tests
model-bakery==1.10.1
pytest==7.2.2
pytest-mock==3.10.0
pytest-django==4.5.2
pytest-pylint==0.19.0
pytest-freezegun==0.4.2
